+++ src/Tests/CuTest.h
@@ -66,7 +66,7 @@ void CuAssertIntEquals_LineMsg(CuTest* tc,
 	int expected, int actual);
 void CuAssertDblEquals_LineMsg(CuTest* tc, 
 	const char* file, int line, const char* message, 
-	double expected, double actual, double delta);
+	long double expected, long double actual, long double delta);
 void CuAssertPtrEquals_LineMsg(CuTest* tc, 
 	const char* file, int line, const char* message, 
 	void* expected, void* actual);
