+++ src/collisions.h
@@ -21,46 +21,46 @@
 /**
     Administration
     */
-void collide(level_data *level, short int who, short int with, bool with_movable, double dt);
+void collide(level_data *level, short int who, short int with, bool with_movable, long double dt);
 void coll_clear_trash(level_data *level, coll_node *node, coll_node *nil);
-void move_objects(level_data *level, double t);
+void move_objects(level_data *level, long double t);
 void collision_min_for_object(level_data *level, short int who);
 void find_next_collision(level_data *level, short int index,
-                         fast_read_set *primitive_done, fast_read_set *movable_done, double time_passed);
+                         fast_read_set *primitive_done, fast_read_set *movable_done, long double time_passed);
 collision_data get_collision_with_primitive(movable_object *who, primitive_object *with_what);
 collision_data get_collision_with_movable(movable_object *who, movable_object *with_whom);
 
-void get_and_check_mov_coll_if_valid(level_data *level, short int who, short int with, double time_passed);
+void get_and_check_mov_coll_if_valid(level_data *level, short int who, short int with, long double time_passed);
 void for_each_higher_check_collision(level_data *level, fast_read_set *movable_done,
-                                     short int who, RB_node *node, RB_node *nil, double time_passed);
+                                     short int who, RB_node *node, RB_node *nil, long double time_passed);
 void in_order_check_collision(level_data *level, fast_read_set *movable_done,
-                              short int who, RB_node *node, RB_node *nil, double time_passed);
+                              short int who, RB_node *node, RB_node *nil, long double time_passed);
 
 /**
     Detecting
     */
-double check_collision_between_two_balls(double x, double y, double dx, double dy,
-                                         double d, double dr);
-double check_collision_between_ball_and_segment(double x, double y, double dx, double dy, double r, double dr, segment *seg);
-double check_exit(double dx, double dy, fixed_exit *ex);
+long double check_collision_between_two_balls(long double x, long double y, long double dx, long double dy,
+                                         long double d, long double dr);
+long double check_collision_between_ball_and_segment(long double x, long double y, long double dx, long double dy, long double r, long double dr, segment *seg);
+long double check_exit(long double dx, long double dy, fixed_exit *ex);
 
 /**
     Colliding
     */
-void get_velocities_after_two_balls_collision(double *v1x, double *v1y, double *v2x, double *v2y,
-                                              double dx, double dy, double m1, double m2, double dr, double restitution);
-void get_velocity_after_ball_to_fixed_ball_collision(double *vx, double *vy, double dx, double dy, double dr, double restitution);
+void get_velocities_after_two_balls_collision(long double *v1x, long double *v1y, long double *v2x, long double *v2y,
+                                              long double dx, long double dy, long double m1, long double m2, long double dr, long double restitution);
+void get_velocity_after_ball_to_fixed_ball_collision(long double *vx, long double *vy, long double dx, long double dy, long double dr, long double restitution);
 
-void get_velocity_after_ball_to_wall_collision(double *vx, double *vy, segment *seg, double dr, double restitution);
+void get_velocity_after_ball_to_wall_collision(long double *vx, long double *vy, segment *seg, long double dr, long double restitution);
 
 /**
     Separating
     */
-bool separate_two_balls(double *x1, double *y1, double m1, double *x2, double *y2, double m2, double d);
-bool separate_ball_from_fixed_ball(double *x1, double *y1, double x2, double y2, double d);
-bool separate_ball_from_segment(double *x, double *y, double d, segment *seg);
+bool separate_two_balls(long double *x1, long double *y1, long double m1, long double *x2, long double *y2, long double m2, long double d);
+bool separate_ball_from_fixed_ball(long double *x1, long double *y1, long double x2, long double y2, long double d);
+bool separate_ball_from_segment(long double *x, long double *y, long double d, segment *seg);
 
 
-void check_dx_dy(double *dx, double *dy);
+void check_dx_dy(long double *dx, long double *dy);
 
 #endif
